Move jest tests to spec/jest
