+++ package.json
@@ -18,6 +18,7 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8.4.33",
+    "postcss-import": "^16.0.0",
     "sass": "^1.69.5",
     "system-font-css": "^2.0.2",
     "tailwindcss": "^3.4.0",
@@ -30,7 +31,7 @@
   },
   "scripts": {
     "build": "bin/vite build",
-    "test": "jest -c app/javascript/jest/jest.config.js",
+    "test": "jest -c spec/jest/jest.config.js",
     "prepare": "husky install",
     "build:css": "tailwindcss -c config/tailwind.config.js -i ./app/assets/stylesheets/application.tailwind.css -o ./app/assets/builds/application.css --minify"
   }
